+++ src/connectedvmware/azext_connectedvmware/vendored_sdks/connectedvmware/operations/_virtual_networks_operations.py
@@ -265,7 +265,7 @@ class VirtualNetworksOperations(object):
     instantiates it for you and attaches it as an attribute.
 
     :ivar models: Alias to model classes used in this operation group.
-    :type models: ~azure_arc_vmware_management_service_api.models
+    :type models: ~azure.mgmt.connectedvmware.models
     :param client: Client for service requests.
     :param config: Configuration of service client.
     :param serializer: An object model serializer.
@@ -357,7 +357,7 @@ class VirtualNetworksOperations(object):
         :param virtual_network_name: Name of the virtual network resource.
         :type virtual_network_name: str
         :param body: Request payload.
-        :type body: ~azure_arc_vmware_management_service_api.models.VirtualNetwork
+        :type body: ~azure.mgmt.connectedvmware.models.VirtualNetwork
         :keyword callable cls: A custom type or function that will be passed the direct response
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
         :keyword polling: By default, your polling method will be ARMPolling. Pass in False for this
@@ -368,8 +368,7 @@ class VirtualNetworksOperations(object):
          Retry-After header is present.
         :return: An instance of LROPoller that returns either VirtualNetwork or the result of
          cls(response)
-        :rtype:
-         ~azure.core.polling.LROPoller[~azure_arc_vmware_management_service_api.models.VirtualNetwork]
+        :rtype: ~azure.core.polling.LROPoller[~azure.mgmt.connectedvmware.models.VirtualNetwork]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         api_version = kwargs.pop('api_version', "2023-03-01-preview")  # type: str
@@ -433,7 +432,7 @@ class VirtualNetworksOperations(object):
         :type virtual_network_name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: VirtualNetwork, or the result of cls(response)
-        :rtype: ~azure_arc_vmware_management_service_api.models.VirtualNetwork
+        :rtype: ~azure.mgmt.connectedvmware.models.VirtualNetwork
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.VirtualNetwork"]
@@ -495,10 +494,10 @@ class VirtualNetworksOperations(object):
         :param virtual_network_name: Name of the virtual network resource.
         :type virtual_network_name: str
         :param body: Resource properties to update.
-        :type body: ~azure_arc_vmware_management_service_api.models.ResourcePatch
+        :type body: ~azure.mgmt.connectedvmware.models.ResourcePatch
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: VirtualNetwork, or the result of cls(response)
-        :rtype: ~azure_arc_vmware_management_service_api.models.VirtualNetwork
+        :rtype: ~azure.mgmt.connectedvmware.models.VirtualNetwork
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.VirtualNetwork"]
@@ -588,8 +587,13 @@ class VirtualNetworksOperations(object):
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             raise HttpResponseError(response=response, error_format=ARMErrorFormat)
 
+        response_headers = {}
+        if response.status_code == 202:
+            response_headers['Location']=self._deserialize('str', response.headers.get('Location'))
+            
+
         if cls:
-            return cls(pipeline_response, None, {})
+            return cls(pipeline_response, None, response_headers)
 
     _delete_initial.metadata = {'url': "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ConnectedVMwarevSphere/virtualNetworks/{virtualNetworkName}"}  # type: ignore
 
@@ -675,8 +679,7 @@ class VirtualNetworksOperations(object):
 
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: An iterator like instance of either VirtualNetworksList or the result of cls(response)
-        :rtype:
-         ~azure.core.paging.ItemPaged[~azure_arc_vmware_management_service_api.models.VirtualNetworksList]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.mgmt.connectedvmware.models.VirtualNetworksList]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         api_version = kwargs.pop('api_version', "2023-03-01-preview")  # type: str
@@ -754,8 +757,7 @@ class VirtualNetworksOperations(object):
         :type resource_group_name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: An iterator like instance of either VirtualNetworksList or the result of cls(response)
-        :rtype:
-         ~azure.core.paging.ItemPaged[~azure_arc_vmware_management_service_api.models.VirtualNetworksList]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.mgmt.connectedvmware.models.VirtualNetworksList]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         api_version = kwargs.pop('api_version', "2023-03-01-preview")  # type: str