+++ src/containerapp/azext_containerapp/_utils.py
@@ -50,7 +50,7 @@ from ._models import (ContainerAppCustomDomainEnvelope as ContainerAppCustomDoma
                       ManagedCertificateEnvelop as ManagedCertificateEnvelopModel,
                       ServiceConnector as ServiceConnectorModel)
 from ._models import OryxMarinerRunImgTagProperty
-from ._managed_service_utils import ManagedRedisUtils, ManagedCosmosDBUtils, ManagedPostgreSQLUtils
+from ._managed_service_utils import ManagedRedisUtils, ManagedCosmosDBUtils, ManagedPostgreSQLFlexibleUtils, ManagedMySQLFlexibleUtils
 
 
 class AppType(Enum):
@@ -439,9 +439,14 @@ def process_service(cmd, resource_list, service_name, arg_dict, subscription_id,
                                                                               name, binding_name))
             elif service["type"] == "Microsoft.DBforPostgreSQL/flexibleServers":
                 service_connector_def_list.append(
-                    ManagedPostgreSQLUtils.build_postgresql_service_connector_def(subscription_id, resource_group_name,
-                                                                                  service_name, arg_dict,
-                                                                                  name, binding_name))
+                    ManagedPostgreSQLFlexibleUtils.build_postgresql_service_connector_def(subscription_id, resource_group_name,
+                                                                                          service_name, arg_dict,
+                                                                                          name, binding_name))
+            elif service["type"] == "Microsoft.DBforMySQL/flexibleServers":
+                service_connector_def_list.append(
+                    ManagedMySQLFlexibleUtils.build_mysql_service_connector_def(subscription_id, resource_group_name,
+                                                                                service_name, arg_dict,
+                                                                                name, binding_name))
             elif service["type"] == "Microsoft.App/containerApps":
                 containerapp_def = ContainerAppClient.show(cmd=cmd, resource_group_name=resource_group_name,
                                                            name=service_name)
@@ -538,7 +543,10 @@ def parse_service_bindings(cmd, service_bindings_list, resource_group_name, name
 
         if not validate_binding_name(binding_name):
             raise InvalidArgumentValueError("The Binding Name can only contain letters, numbers (0-9), periods ('.'), "
-                                            "and underscores ('_'). The length must not be more than 60 characters.")
+                                            "and underscores ('_'). The length must not be more than 60 characters. "
+                                            "By default, the binding name is the same as the service name you specified "
+                                            "[my-aca-pgaddon], but you can override the default and specify your own "
+                                            "compliant binding name like this --bind my-aca-pgaddon[:my_aca_pgaddon].")
 
         resource_client = get_mgmt_service_client(cmd.cli_ctx, ResourceManagementClient)
 
@@ -600,14 +608,6 @@ def parse_auth_flags(auth_list):
         })
 
     return auth_def
-
-
-def _update_revision_env_secretrefs(containers, name):
-    for container in containers:
-        if "env" in container:
-            for var in container["env"]:
-                if "secretRef" in var:
-                    var["secretRef"] = var["secretRef"].replace("{}-".format(name), "")
 
 
 def _update_revision_env_secretrefs(containers, name):