+++ src/dataprotection/azext_dataprotection/aaz/latest/dataprotection/backup_vault/_create.py
@@ -26,9 +26,9 @@ class Create(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2023-01-01",
+        "version": "2023-05-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.dataprotection/backupvaults/{}", "2023-01-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.dataprotection/backupvaults/{}", "2023-05-01"],
         ]
     }
 
@@ -155,11 +155,13 @@ class Create(AAZCommand):
             options=["--soft-delete-retention", "--retention-duration-in-days"],
             arg_group="SoftDeleteSettings",
             help="Soft delete retention duration",
+            default=14.0,
         )
         _args_schema.soft_delete_state = AAZStrArg(
             options=["--soft-delete-state"],
             arg_group="SoftDeleteSettings",
             help="State of soft delete",
+            default="On",
             enum={"AlwaysOn": "AlwaysOn", "Off": "Off", "On": "On"},
         )
         return cls._args_schema
@@ -245,7 +247,7 @@ class Create(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-01-01",
+                    "api-version", "2023-05-01",
                     required=True,
                 ),
             }
@@ -385,6 +387,22 @@ class Create(AAZCommand):
                 flags={"read_only": True},
             )
             identity.type = AAZStrType()
+            identity.user_assigned_identities = AAZDictType(
+                serialized_name="userAssignedIdentities",
+            )
+
+            user_assigned_identities = cls._schema_on_200_201.identity.user_assigned_identities
+            user_assigned_identities.Element = AAZObjectType()
+
+            _element = cls._schema_on_200_201.identity.user_assigned_identities.Element
+            _element.client_id = AAZStrType(
+                serialized_name="clientId",
+                flags={"read_only": True},
+            )
+            _element.principal_id = AAZStrType(
+                serialized_name="principalId",
+                flags={"read_only": True},
+            )
 
             properties = cls._schema_on_200_201.properties
             properties.feature_settings = AAZObjectType(
@@ -408,6 +426,10 @@ class Create(AAZCommand):
                 serialized_name="resourceMoveState",
                 flags={"read_only": True},
             )
+            properties.secure_score = AAZStrType(
+                serialized_name="secureScore",
+                flags={"read_only": True},
+            )
             properties.security_settings = AAZObjectType(
                 serialized_name="securitySettings",
             )
@@ -417,9 +439,15 @@ class Create(AAZCommand):
             )
 
             feature_settings = cls._schema_on_200_201.properties.feature_settings
+            feature_settings.cross_region_restore_settings = AAZObjectType(
+                serialized_name="crossRegionRestoreSettings",
+            )
             feature_settings.cross_subscription_restore_settings = AAZObjectType(
                 serialized_name="crossSubscriptionRestoreSettings",
             )
+
+            cross_region_restore_settings = cls._schema_on_200_201.properties.feature_settings.cross_region_restore_settings
+            cross_region_restore_settings.state = AAZStrType()
 
             cross_subscription_restore_settings = cls._schema_on_200_201.properties.feature_settings.cross_subscription_restore_settings
             cross_subscription_restore_settings.state = AAZStrType()