+++ src/managednetworkfabric/azext_managednetworkfabric/tests/latest/test_ipextendedcommunity.py
@@ -26,6 +26,7 @@ def call_scenario1(test):
     step_create(test, checks=[])
     step_show(test, checks=[])
     step_list_resource_group(test, checks=[])
+    step_list_subscription(test, checks=[])
     step_delete(test, checks=[])
     cleanup_scenario1(test)
 
@@ -47,6 +48,12 @@ def step_list_resource_group(test, checks=None):
     if checks is None:
         checks = []
     test.cmd('az networkfabric ipextendedcommunity list --resource-group {rg}')
+
+def step_list_subscription(test, checks=None):
+    '''ipextendedcommunity list by subscription'''
+    if checks is None:
+        checks = []
+    test.cmd('az networkfabric ipextendedcommunity list')
 
 def step_delete(test, checks=None):
     '''ipextendedcommunity delete operation'''