+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/cluster/baremetalmachinekeyset/_list.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud cluster baremetalmachinekeyset list",
-    is_preview=True,
 )
 class List(AAZCommand):
     """List bare metal machine key sets of the cluster.