+++ src/spring/HISTORY.md
@@ -1,8 +1,16 @@
 Release History
 ===============
-1.14.2
+1.14.4
 ---
 * Add arguments `--type` and `--git-sub-path` in `spring application-accelerator customized-accelerator create` and `spring application-accelerator customized-accelerator update` for accelerator fragment support.
+
+1.14.3
+---
+* Make error message for `az spring app logs` more readable.  
+
+1.14.2
+---
+* Add new command `az spring flush-virtualnetwork-dns-settings` to flush virtual network DNS settings for the service instance.
 
 1.14.1
 ---